rename zk_proofs_screen to grovestark_screen
