+++ src/ui/tools/mod.rs
@@ -5,4 +5,4 @@ pub mod platform_info_screen;
 pub mod proof_log_screen;
 pub mod proof_visualizer_screen;
 pub mod transition_visualizer_screen;
-pub mod zk_proofs_screen;
+pub mod grovestark_screen;
